Theme changes to take account of Hugo 0.18
There were breaking changes in Hugo 0.18 which exposed poor theme
organisation and hacks used in squarepenguin.co.uk.

Specifically, section generation templates MUST now be in the correct
hierarchy in the correct directories. For example, previously the
guides.html page in content directory governed the appearence of the final
/guides/ web page. Now guides.html in that directory is not recognised and
it's content
along with restructured layout MUST go in the /layouts/section directory
under the name guides.html and this will govern the page layout for the
/guides/ top level section page.

Web pages at /guides/pages-here/ are still themed by layouts in
/layouts/guides/

Further work will be needed to take account for Hugo's new 'everything is
a page' approach once their docs are updated.